Check can now be broken by capturing
The program had a bug that a checkmate was interpreted when there is
no safe squares for the king to move to or any way to put a piece
between the king and the attacker.
Before the program did not check capturing the attacking piece as a
possibility to escape check.